Add Model and ModelTest projects
Right now these are just Xcode's default macOS static library and
unit test templates. First goal: parse a sample JSON file and check
that the output looks correct.

I'm using a macOS project so that I don't have to muck about with
mobile devices right now. This code should all be portable.

If we wanted to target both iOS and Android, I might try to make
this completely cross-platform, say by writing it in C++, but I
think that's beyond the scope of this little exercise.