Add Player.imageUrl property
I realised that trying to handle asynchronous loading in the
Players and Player classes was making them do two different
things -- they just represent the data, they shouldn't have
to worry about how it get loaded.

Therefore, Player just has an imageUrl property and it's up
to clients to figure out how to load it.

A bit of refactoring to handle relative URLs correctly (which
I think *is* core functionality for these classes). Finally
switched from NSString to NSURL to make this easier.