Update SharedPlayer to use the environment instead of a static global
I learned that a static variable isn't a good way to inject data
into an app under test. :) It gets reset at inconvenient times.

This approach works better, as the app environment is explicitly
controlled and programmable. Unfortunately the test itself has a
separate environment from the app under test, so this unit test
has to use setenv(), which feels a bit hacky. But it works.